user/requirements.txt
Django==3.2
djangorestframework
requests
django-environ
